Increasing House quantities and Matrix Size
Worked most of my work on test2.py with few edits to test.py and none to pizza.py or the main files. Also included a reference image of the map of houses and their connections with distances shown. In test2.py I was able to successfully get the algorithm to traverse an increased matrix with more houses and more connections. Found a bug when trying to traverse from Z - B of repeating connections, plan is to implement this to the main file to handle more houses and to soon implement the concept of gas consumption too