A Far Jump
Removed excess files that were uploaded for testing. Updated Pizza.py to have the current state of the algorithm. Works kind of how I want it and it got turned it for the final. However I want to rework it a lot, clean it up and make it more mutable, and add a GUI for it. A lot to come., and a brief explaining of the project in Final Project Report PDF